+++ packages/kinvey-nativescript-sdk/webpack.config.android.js
@@ -32,6 +32,7 @@ module.exports = (env = {}) => {
       'application': 'application',
       'tns-core-modules/application': 'application',
       http: 'http',
+      './kinvey-nativescript-sdk': './kinvey-nativescript-sdk',
       'tns-core-modules/http': 'http',
       'tns-core-modules/file-system': 'file-system',
       'tns-core-modules/ui/frame': 'ui/frame',
@@ -67,7 +68,7 @@ module.exports = (env = {}) => {
   bundleName = `${bundleName}.${platform}`;
   pushBundleName = `${pushBundleName}.${platform}`;
   config.entry[bundleName] = './src/index.ts';
-  config.entry[pushBundleName] = './src/push.ts';
+  config.entry[pushBundleName] = './src/push';
   return config;
 };
 
@@ -125,7 +126,8 @@ function getPlugins(env, platform) {
         }
       },
       { from: '.travis.yml' },
-      { from: path.join(__dirname, '../../src/kinvey.d.ts') },
+      { from: path.join(__dirname, 'kinvey.d.ts') },
+      { from: path.join(__dirname, 'push.d.ts') },
       { from: 'platforms/android/**/*' },
       { from: 'LICENSE' },
       { from: 'README.md' },